Fix query2 parser star select issue
Variable were parsed by the parser class but never added to the query
object. Projection vars are now added to the object thus fixing the
resulting * queries.

This fix introduced another issue which was also fixed. When no
visible graph is found, the store invalidates the query by settings the
where clause to a false filter. Virtuoso does not like such queries
that have a valist in the select clause. It crashes with a out of mem
error. This was fixed by also settings the varlist to * again when the
query where clause gets falsified.