add option to export SPARQL result to CSV
This adds an option to the Virtuoso adapter to export queries as
CSV. For better structuring, the CSV converter should be moved up a
level to be called from the Erfurt/Store.php. Right now, there is not
much of output format handling there, but it would probably be best to
check for supported formats (as suggested by the TODO in line 1565) and
use either a default converter or (if supported) the custom (and
hopefully faster) method of the used store.