More seamless integration of intuitive search
After switching to AND as default in the filter (see commit
bf7455dbd82b656856e2b834cd9765e894fb6a83), this commit makes it possible
to mix phrases and words (e.g. "Journal 'Chemical Theory'" will be translated
to "'Journal' AND 'Chemical Theory'). As soon as one of the keywords AND, OR
or NEAR is used, it is assumed the input string adheres to the bif:contains
syntax and will not be translated.