First stub to enable property paths in queries
Property paths are a SPARQL 1.1 feature, which is supported by Virtuoso
but not via Erfurt, because the query gets invalidated by the parser before
it's sent to Virtuoso. A node in a SPARQL query is normally checked to
determine whether it is IRI, qname, literal, etc... with this commit, a node
is also checked for being a property path. However, said check is right now
only very rudimentary.